+++ Arkane/Arkane/Renderer/Buffers/VertexBufferObject.h
@@ -14,15 +14,15 @@ public:
 	~VertexBufferObject();
 
 	// Allocate or free the buffer.
-	bool AllocBufferObject(const void* _data, uint32_t _allocSize, EBufferUsage _usage);
+	bool AllocBufferObject(const void* _data, size_t _allocSize, EBufferUsage _usage);
 	void FreeBufferObject();
 
 	// Make this buffer a reference to another buffer.
 	void Reference(const VertexBufferObject& _other);
-	void Reference(const VertexBufferObject& _other, uint32_t _refOffset, uint32_t _refSize);
+	void Reference(const VertexBufferObject& _other, size_t _refOffset, size_t _refSize);
 
 	// Copies data to the buffer. 'size' may be less than the originally allocated size.
-	void Update(const void* _data, uint32_t _size, uint32_t _offset = 0) const;
+	void Update(const void* _data, size_t _size, size_t _offset = 0) const;
 
 	void* MapBuffer(EBufferMappingType _mapType);
 	void UnmapBuffer();
